backend/requirements.txt
django
psycopg2
djangorestframework
tqdm